+++ src/main/java/ua/foxminded/pskn/universitycms/repository/university/UniversityRepository.java
@@ -13,7 +13,7 @@ import java.util.Optional;
 public interface UniversityRepository extends JpaRepository<University, Long> {
     Optional<University> findByUniversityName(String universityName);
 
-    boolean existsByUniversityId(int universityId);
+    boolean existsByUniversityId(Long universityId);
 
     @Modifying
     @Query("UPDATE University u SET u.universityName = :name WHERE u.universityId = :id")
